requirements.txt
streamlit
numpy==1.23.5
tensorflow==2.12.0
ml_dtypes==0.2.0
pillow
torch
torchvision
fastapi
uvicorn
python-multipart
